requirements.txt
streamlit>=1.32.0
google-genai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
